Remove `TransactionView` in favour of `TransactionSummaryView`
